# api/swagger.yaml
openapi: 3.0.3
info:
  title: lifeos API
  description: Personal Life OS — Go/Gin + SQLite + JWT, single-user lokal.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: lokal
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security: []
paths:
...
  /v1/workouts:
    get:
      summary: List workouts (?from=&to=)
      operationId: listWorkouts
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: list }
    post:
      summary: Catat workout
      operationId: createWorkout
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, type]
              properties:
                date: { type: string, format: date }
                type: { type: string, example: Lari }
                duration_min: { type: integer }
                intensity: { type: string }
      responses:
        "201": { description: created }
  /v1/learning:
    get:
      summary: List learning
      operationId: listLearn
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: list }
    post:
      summary: Tambah learning
      operationId: createLearn
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic]
              properties:
                topic: { type: string }
                type: { type: string }
                progress: { type: integer }
                hours: { type: number }
      responses:
        "201": { description: created }
  /v1/learning/{id}:
    put:
      summary: Update learning
      operationId: updateLearn
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic]
              properties:
                topic: { type: string }
      responses:
        "200": { description: ok }
    delete:
      summary: Hapus learning
      operationId: deleteLearn
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: deleted }
  /v1/reading:
    get:
      summary: List reading
      operationId: listRead
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: list }
    post:
      summary: Tambah reading
      operationId: createRead
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                type: { type: string }
                status: { type: string }
      responses:
        "201": { description: created }
  /v1/reading/{id}:
    put:
      summary: Update reading
      operationId: updateRead
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
      responses:
        "200": { description: ok }
    delete:
      summary: Hapus reading
      operationId: deleteRead
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: deleted }
...
  /v1/reviews/{id}:
    get:
      summary: Detail review
      operationId: getReview
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: ok }
        "404": { description: not_found }
    put:
      summary: Update review (stats hitung ulang)
      operationId: updateReview
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [week_start]
              properties:
                week_start: { type: string, format: date }
      responses:
        "200": { description: ok }
    delete:
      summary: Hapus review
      operationId: deleteReview
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: deleted }
...